requirements.txt
hitran-api>=1.1.2.0
SQLAlchemy>=1.2.7
numpy>=1.17.5+mkl
numba>=0.41.0
tabulate>=0.8.2

python-dateutil>=2.8.1
pyparsing>=2.2.0
scipy>=1.0.1
matplotlib>=2.2.2
jupyter>=1.0.0